+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.23.0
 
+* Adds device data to transactions, customers, and credit cards.
 * Lots of cleanups [thanks to https://github.com/steve-nester-uk]
 
 ## 2.22.1